+++ src/plugins/external_tool_support/src/bowtie/bowtie_tests/bowtieTests.cpp
@@ -270,7 +270,7 @@ void GTest_Bowtie::cleanup()
 {
 
     // delete index
-    if(!usePrebuildIndex) {
+    if(!hasError() && !usePrebuildIndex) {
         QString prefix = env->getVar("TEMP_DATA_DIR")+"/"+QString::number(getTaskId());
         QStringList files(QStringList() << prefix+".1.ebwt" << prefix+".2.ebwt"
             << prefix+".3.ebwt" << prefix+".4.ebwt" << prefix+".rev.1.ebwt" << prefix+".rev.2.ebwt");
@@ -284,10 +284,12 @@ void GTest_Bowtie::cleanup()
     }
     //delete tmp result
     QFileInfo tmpResult(config.resultFileName.getURLString());
-    if (tmpResult.exists()) {
+    if (!hasError() && tmpResult.exists()) {
         ioLog.trace(QString("Deleting tmp result file :%1").arg(tmpResult.absoluteFilePath()));
         QFile::remove(tmpResult.absoluteFilePath());
     }
+
+    XmlTest::cleanup();
 }
 
 GTest_Bowtie::~GTest_Bowtie()
